requirements.txt
aenum==3.1.15
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.4
h11==0.14.0
httplib2==0.22.0
idna==3.10
load-dotenv==0.1.0
loguru==0.7.2
# Editable Git install with no remote (njema==0.0.1)
-e /home/kanairo/Private/cashathand/Njemz/Njema
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
